requirements.txt
# ORBIT Python ADK Implementation Requirements
# Core dependencies for the ORBIT image processing workflow system

# Google Agent Development Kit
google-adk>=1.0.0

# FastMCP for Model Context Protocol server
fastmcp>=0.1.0

# HTTP client library
httpx>=0.27.0

# Async utilities
asyncio-mqtt>=0.16.0

# Environment variables
python-dotenv>=1.0.0

# JSON handling and validation
pydantic>=2.0.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Date and time utilities
python-dateutil>=2.8.0

# Logging utilities
structlog>=23.0.0

# Type checking (development)
mypy>=1.0.0
types-requests>=2.31.0

# Code formatting (development)
black>=23.0.0
isort>=5.12.0

# Development dependencies
pytest-mock>=3.11.0
pytest-cov>=4.1.0